添加control测试类，修正dto/containner/containnerlist属性类型

+++ src/main/java/com/autodl_backend/autodl/dto/container/ContainerListItem.java
@@ -38,7 +38,7 @@ public class ContainerListItem {
     private Integer cpuNum;
 
     @JsonProperty("memory_size")
-    private Integer memorySize;
+    private Long memorySize;
 
     @JsonProperty("image_uuid")
     private String imageUuid;
